requirements.txt
anyio==3.6.2
boto3==1.26.129
botocore==1.29.129
click==8.1.3
fastapi==0.95.1
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
strawberry-graphql==0.176.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
